Replace string filename with filesystem path in compiler
In this commit, we have replaced the standard string filename with a filesystem path in the `compile` function signature and also updated associated function calls. This change provides better support for filesystem operations and enhances the type safety, as it restricts the type of input to be a valid filesystem path. This way, we can avoid any misuse or errors where a string that does not represent a valid filesystem path is input to the `compile` function.

Further, the initialization of builtin compiler has been moved to a separate thread to optimize the execution time and manage other tasks simultaneously.