Add Vector and String methods with UInt32 index argument
This commit adds new overloads of "__at__" methods to Vector and String classes that accept UInt32 index argument in the builtin_functions.cpp. This change broadens the range of valid inputs for accessing elements of Vector and String objects, extending the compatibility with user code. This addition is a response to reports of UInt32 being a common type for indices in consumer codebases, helping to avoid unnecessary conversions.